Revert "Overhaul"
