--- loadshedding/logs/another.log
@@ -1,1 +1,0 @@
-21:05:26.679 [main] ERROR load_shedding_sim.LoadSheddingSimMain - Did it again!
