--- loadshedding/logs/app.log
@@ -1,2 +1,0 @@
-21:05:10.348 [main] ERROR load_shedding_sim.LoadSheddingSimMain - Did it again!
-21:05:26.679 [main] ERROR load_shedding_sim.LoadSheddingSimMain - Did it again!
