+++ loadshedding/src/main/java/load_shedding_sim/LoadSheddingSimMain.java
@@ -2,6 +2,7 @@ package load_shedding_sim;
 
 import org.apache.logging.log4j.*;
 
+
 import java.text.SimpleDateFormat;
 import java.util.*;
 import java.io.*;
@@ -13,8 +14,6 @@ import org.apache.commons.lang3.time.DateUtils;
 import com.google.common.base.Charsets;
 import com.google.common.io.Files;
 
-import data_generation.AnalysisInput;
-
 
 public class LoadSheddingSimMain {
 	public static void main (String[] args) throws Exception{
@@ -25,9 +24,7 @@ public class LoadSheddingSimMain {
 		//OracleMakingVistaTV myOracle = new OracleMakingVistaTV(Debug.ORACLE_LOG);
 		//myOracle.making();
 		
-//		AnalysisInput myAnalysis = new AnalysisInput();
-//		myAnalysis.singleOracle();
-//		
+		//AnalysisInput myAnalysis = new AnalysisInput();
 		//myAnalysis.getHistogram();
 		//myAnalysis.getHistogramWithLocalSimTime();
