+++ loadshedding/src/main/java/strategy/DataCacheClock.java
@@ -1,4 +1,4 @@
-package strategies;
+package strategy;
 import java.util.*;
 
 import data_entry.DataEntry;
@@ -30,7 +30,6 @@ public class DataCacheClock extends DataCache{
 		this.statOfTotalEvication++;
 		tempItr.set(input);
 		this.deleteFromStore(tempData);
-		index.remove(tempData);
 		//store.remove(tempData.key, tempData);
 		pointer = tempItr.nextIndex();
 		return tempData;
@@ -43,7 +42,7 @@ public class DataCacheClock extends DataCache{
 	public void insertOneEntry ( DataEntry newEntry) {
 		if( store.size() >= allowedSize ) {
 			replaceVictimEntry(newEntry);
-			index.add(newEntry);
+			//System.out.println(this.index.size());
 		} else
 			index.add(newEntry);
 		this.putintoStore(newEntry);
@@ -51,11 +50,22 @@ public class DataCacheClock extends DataCache{
 	public void garbageCollection (Date currentSystemReadTimeStamp) {
 		if(this.enableReasoning) {
 			DataEntry temEntry = this.endingTimeQ.peek();
+			//if(temEntry != null)
+			//	System.out.println(Debug.sdf.format(currentSystemReadTimeStamp) +"\t"+ Debug.sdf.format(temEntry.timeStampEnd)+"\t"+ this.isInner);
 			while (temEntry != null && temEntry.timeStampEnd.before(currentSystemReadTimeStamp)) {
+				
+//				System.out.println("deleting"+ Debug.sdf.format(currentSystemReadTimeStamp) +"\t"+ Debug.sdf.format(temEntry.timeStampEnd)+"\t"+ this.isInner);
+//				System.out.println(this.allowedSize);
+//				System.out.println(this.store.size());
+//				System.out.println(this.index.size());
+				
 				this.deleteFromStore(temEntry);
 				this.statOfTotalExpiried++;
 				index.remove(temEntry);
 				temEntry = this.endingTimeQ.peek();
+				
+//				System.out.println(this.store.size());
+//				System.out.println(this.index.size());
 			}
 		}
 	}
