+++ loadshedding/src/main/java/strategy/DataCacheFIFOFIFO.java
@@ -1,4 +1,4 @@
-package strategies;
+package strategy;
 import java.util.*;
 
 import load_shedding_sim.Debug;
