+++ loadshedding/src/main/java/strategy/LRUCache.java
@@ -1,4 +1,4 @@
-package strategies;
+package strategy;
 
 /*
  * XAPool: Open Source XA JDBC Pool
@@ -105,8 +105,10 @@ public class LRUCache
                 last = node.prev;
             if (first == node)
                 first = node.next;
+            
             currentSize--;
         }
+        
         return node.value;
     }
 
@@ -120,13 +122,11 @@ public class LRUCache
     {
         if(last != null)
         {
-        	CacheNode tempNode = last;
             if(last.prev != null)
                 last.prev.next = null;
             else
                 first = null;
             last = last.prev;
-            tempNode = null;
         }
     }
     public DataEntry removeLastEntry() {
