--- loadshedding/src/main/resources/log4j2.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="WARN">
-  <Appenders>
-    <Console name="Console" target="SYSTEM_OUT">
-      <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level: %logger{36} %M - %msg%n"/>
-    </Console>
-     
-    <File name="MyFile" fileName="logs/app.log">
-        <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n"/>
-    </File> 
-    <File name="AnotherFile" fileName="logs/another.log">
-        <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n"/>
-    </File>  
-    
-  </Appenders>
-  <Loggers>
-  	<logger name="ARC" level="info"/>
-    <Root level="all">
-      <AppenderRef ref="Console" level="all"/>
-    </Root>
-  </Loggers>
-</Configuration>
